requirements.txt
Django==3.1.12
djangorestframework==3.12.4
djongo==1.3.7
pymongo==3.11.4
django-cors-headers==3.7.0  
face_recognition
deep-translator
gTTS
pygame
tensorflow==2.19.0